Update reset password & confirm page
